requirements.txt
termcolor
langgraph
langchain-community
arxiv
wikipedia
duckduckgo-search
aiohttp
pygraphviz
requests
pyvmomi